requirements.txt
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
et-xmlfile==1.1.0
executing==2.0.0
feature-engine==1.4.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.2.1
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.0
h11==0.14.0
html5lib==1.1
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lightgbm==4.1.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
openpyxl==3.1.2
outcome==1.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
PyGithub==2.1.1
Pygments==2.16.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
selenium==4.14.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.22
stack-data==0.6.3
statsmodels==0.14.0
stripe==7.0.0
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.1.0
xgboost==2.0.1